rename conda/core/package_cache.py to conda/core/package_cache_data.py
