+++ conda/misc.py
@@ -17,7 +17,7 @@ from .common.url import is_url, join_url, path_to_url, unquote
 from .core.index import get_index
 from .core.link import PrefixSetup, UnlinkLinkTransaction
 from .core.linked_data import PrefixData, linked_data
-from .core.package_cache import PackageCacheData, ProgressiveFetchExtract
+from .core.package_cache_data import PackageCacheData, ProgressiveFetchExtract
 from .exceptions import PackagesNotFoundError, ParseError
 from .gateways.disk.delete import rm_rf
 from .gateways.disk.link import islink, symlink, readlink
